+++ samples/sample-client-common/src/main/java/com/flipkart/aesop/sample/client/processor/ConsoleProcessor.java
@@ -1,0 +1,21 @@
+package com.flipkart.aesop.sample.client.processor;
+
+import com.flipkart.aesop.event.AbstractEvent;
+import com.flipkart.aesop.processor.DestinationEventProcessor;
+import org.trpr.platform.core.impl.logging.LogFactory;
+import org.trpr.platform.core.spi.logging.Logger;
+
+import javax.naming.OperationNotSupportedException;
+
+/**
+ * Entity Upsert Data Layer. Persists {@link com.linkedin.databus.core.DbusOpcode#UPSERT} events to Logs.
+ * @author arya.ketan
+ */
+public class ConsoleProcessor implements DestinationEventProcessor
+{
+    public static final Logger LOGGER = LogFactory.getLogger(ConsoleProcessor.class);
+    public void processDestinationEvent(AbstractEvent abstractEvent) throws OperationNotSupportedException
+    {
+        LOGGER.info("CONSOLE_PROCESSOR:Event Type: " + abstractEvent.getEventType().toString() + " Event : " + abstractEvent.toString());
+    }
+}